app/requirements.txt
fastapi
uvicorn[standard]
SQLAlchemy
alembic
passlib[bcrypt]
bcrypt==4.0.1
psycopg2-binary
python-jose[cryptography]
python-dotenv
python-multipart
pydantic-settings
pydantic[email]